==> 06a_Testes_Connectivity.rmd <==
---
title: "06a_Testes_Module_Trait_Relationship.rmd"
author: "Garett Slater"
date: "April 21 2022"
output:
  html_document: default
  pdf_document: default
---


```{r knitr setup, include=FALSE,  eval=TRUE, echo=FALSE, warning=FALSE}
library(knitr)
knitr::opts_chunk$set(eval=TRUE, cache=FALSE, message=FALSE, warning=FALSE, 
                      comment = "")
```


```{r echo=FALSE}
library(WGCNA)
setwd("~/OneDrive/Manuscripts/GenesofDrones/WGCNA")
load(file="Testes-networkConstruction-auto.RData")
load(file="T_datExpr.RData")
load(file="T_adj.RData")
```


# Top Hub genes for each cluster
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/GenesofDrones/WGCNA")

# Top hub genes for each module:
chooseTopHubInEachModule(
  T_datExpr, 
  T_moduleColors, 
  omitColors = "grey", 
  power = 6, 
  type = "unsigned")
```

- kTotal: the whole network connectivity (adjacency). Connectivity of the each gene based on its r-values to all other genes in the whole network
- kWithin: the within module connectivity (intramodular connectivity). Connectivity of the each gene within a single module based on its r-values to all other genes within the same module
- kOut=kTotal-kWithin
- kDiff=kIn-kOut=2*kIN-kTotal

```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/GenesofDrones/WGCNA")
#Representing modules by eigengenes and relating eigengenes to one another
##To  get  a  sense  of  how  related  the  modules  are  one  can  summarize  each  module  
#by  its  eigengene  (first  principal component).

#Calculates module eigengenes (1st principal component) of modules in a given single dataset.
T_datME = moduleEigengenes(T_datExpr, T_moduleLabels, softPower = 6)$eigengenes

# significant digits (signif) to be used.
datME_T_table = signif(cor(T_datME, use="p"), 2)
write.csv(datME_T_table, file = "datME_T_table.csv")


# Intramodular connectivity
# Calculating the intramodular connectivity for each gene.  (In network literature, connectivity is often referred to as ”degree”)
# The function intramodularConnectivity computes:
#--> kTotal: the whole network connectivity (adjacency). Connectivity of the each gene based on its r-values to all other genes in the whole network
#--> kWithin: the within module connectivity (intramodular connectivity). Connectivity of the each gene within a single module based on its r-values to all other genes within the same module
#--> kOut=kTotal-kWithin
#--> kDiff=kIn-kOut=2*kIN-kTotal


#Calculates intramodular connectivity, i.e., connectivity of nodes to other nodes within the same module.
Alldegrees_T_colors = intramodularConnectivity(T_adj, T_moduleColors)
head(Alldegrees_T_colors)
```

# Generalizing intramodular connectivity for all genes on the array
The intramodular connectivity measure is only defined for the genes inside a given module. 
But in practice it can be very important to measure how connected a given genes is to biologically interesting modules. 
A module eigengene-based connectivity measure for each gene as the correlation between a the gene expression and the module eigengene. In summary, how is the relationship between a gene and any module.
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/GenesofDrones/WGCNA")
#Generalizing intramodular connectivity for all genes on the array
##The intramodular connectivity measure is only defined for the genes inside a given module. 
## But in practice it can be very important to measure how connected a given genes is to biologically interesting modules. 
##A module eigengene-based connectivity measure for each gene as the correlation between a the gene expression and the module eigengene. In summary, how is the relationship between a gene and any module.

#output --> We have a module membership value for each gene in each module.


T_datKME = abs(signedKME(T_datExpr, T_MEs, outputColumnName="MM."))

# Display the first few rows of the data frame

head(T_datKME)
write.csv(T_datKME, file = "MM_T.csv")

```

# Calculates module eigengenes (1st principal component) of modules in a given single dataset.

```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/GenesofDrones/WGCNA")
# Calculates module eigengenes (1st principal component) of modules in a given single dataset.
datME_T_colors = moduleEigengenes(T_datExpr,T_moduleColors, softPower = 6)$eigengenes

signif(cor(datME_T_colors, use="p"), 2)
datKME_T_colors = abs(signedKME(T_datExpr, datME_T_colors, outputColumnName="MM."))
head(datKME_T_colors)

```
